+++ apps/web/app/composables/useUtilityBarConfiguration/index.ts
@@ -1,0 +1,1 @@
+export * from './useUtilityBarConfiguration';
